backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
colorama==0.4.6
email-validator==2.3.0
fastapi==0.128.4
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
psycopg2-binary
python-dotenv
sqlalchemy
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
passlib[argon2]
python-jose[cryptography]
Pillow
